Clear inputs on form submission
