+++ .eslintrc.yaml
@@ -16,6 +16,7 @@ rules:
   react-hooks/rules-of-hooks: error
   react-hooks/exhaustive-deps: warn
   react/prop-types: off
+  '@typescript-eslint/ban-ts-comment': off
 settings:
   react:
     pragma: React
